requirements.txt
flask

google-generativeai
python-dotenv
ffmpeg-python
SpeechRecognition
pyaudio